Refactor retrieval endpoints and dont paginate top agents
